// frontend/src/types/Candle.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Timeframe } from "./Timeframe";

export type Candle = { timestamp: number, exchange: string, symbol: string, timeframe: Timeframe, open: string, high: string, low: string, close: string, volume: string, };